Some changes to unit tests for new Image classes.  And a bug fix found by running tests. (#117)

+++ include/galsim/Image.h
@@ -211,7 +211,7 @@ namespace galsim {
         /**
          *  @brief Set the pixel scale 
          */
-        void setScale(int scale) { _scale = scale; }
+        void setScale(double scale) { _scale = scale; }
 
         /**
          *  @brief Move the center of the image, changing the logical location of the pixels.
